Unified CommentsParser and RobotCommentsParser
