+++ src/test/java/com/urswolfer/gerrit/client/rest/http/changes/ChangesRestClientTest.java
@@ -93,7 +93,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser, null, null,
             null, null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
 
         Changes.QueryRequest queryRequest = changes.query();
         testCase.queryParameter.apply(queryRequest).get();
@@ -111,7 +111,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
         changesRestClient.query("is:open").get();
 
         EasyMock.verify(gerritRestClient);
@@ -125,7 +125,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
 
         ChangeApi changeApi = changesRestClient.id(123);
 
@@ -140,7 +140,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
 
         ChangeApi changeApi = changesRestClient.id("packages/test", 123);
 
@@ -155,7 +155,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null,  null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
 
         ChangeApi changeApi = changesRestClient.id("packages/test", "master", "Ieabd72e73f3da0df90fd6e8cba8f6c5dd7d120df");
 
@@ -170,7 +170,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser,null, null,
             null, null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
 
         changes.query().get();
 
@@ -191,7 +191,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser, null, null,
             null, null, null, null, null,
-            null, null, null, null, null, null);
+            null, null, null, null, null);
 
         ChangeApi changeApi = changes.create(changeInput);
