+++ src/test/java/com/urswolfer/gerrit/client/rest/http/changes/RevisionApiRestClientTest.java
@@ -383,7 +383,6 @@ public class RevisionApiRestClientTest extends AbstractParserTest {
     private ChangesRestClient getChangesRestClient(GerritRestClient gerritRestClient) {
         ChangesParser changesParser = EasyMock.createMock(ChangesParser.class);
         CommentsParser commentsParser = EasyMock.createMock(CommentsParser.class);
-        RobotCommentsParser robotCommentsParser = EasyMock.createMock(RobotCommentsParser.class);
         MessagesParser messagesParser = EasyMock.createMock(MessagesParser.class);
         IncludedInInfoParser includedInInfoParser = EasyMock.createMock(IncludedInInfoParser.class);
         FileInfoParser fileInfoParser = EasyMock.createMock(FileInfoParser.class);
@@ -394,7 +393,7 @@ public class RevisionApiRestClientTest extends AbstractParserTest {
         ReviewResultParser reviewResultParser = EasyMock.createMock(ReviewResultParser.class);
         CommitInfoParser commitInfoParser = EasyMock.createMock(CommitInfoParser.class);
         HashtagsParser hashtagsParser = EasyMock.createMock(HashtagsParser.class);
-        return new ChangesRestClient(gerritRestClient, changesParser, commentsParser, robotCommentsParser, messagesParser,
+        return new ChangesRestClient(gerritRestClient, changesParser, commentsParser, messagesParser,
             includedInInfoParser, fileInfoParser, diffInfoParser, null, reviewerInfoParser, editInfoParser,
             addReviewerResultParser, reviewResultParser, commitInfoParser, hashtagsParser);
     }
@@ -404,7 +403,6 @@ public class RevisionApiRestClientTest extends AbstractParserTest {
                 gerritRestClient,
                 EasyMock.createMock(ChangesParser.class),
                 commentsParser,
-                EasyMock.createMock(RobotCommentsParser.class),
                 EasyMock.createMock(MessagesParser.class),
                 EasyMock.createMock(IncludedInInfoParser.class),
                 EasyMock.createMock(FileInfoParser.class),
